Move main frame init extensions to runnable
